forwarding: extract policy-based-routing override into pbr.rs
Pure code-motion split of forwarding/mod.rs (#5650). Move the
ingress route-table override resolver (ingress_route_table_override) and
its PbrRejectSink / RouteOverride support types into forwarding/pbr.rs.

Functions are moved verbatim; no logic, signature, or hot-path change.
Both the struct declaration AND its pub(super) fields become
pub(in crate::afxdp) so the external constructor in poll_descriptor
(which builds PbrRejectSink { .. }) keeps compiling; the mod.rs glob
re-export keeps all call sites resolving.

Validation: cargo build clean (0 errors).
