+++ userspace-dp/src/afxdp/forwarding/mod.rs
@@ -2,6 +2,8 @@ use super::*;
 use std::borrow::Cow;
 
 mod host_inbound;
+mod pbr;
+pub(in crate::afxdp) use pbr::*;
 mod ipsec;
 pub(in crate::afxdp) use ipsec::*;
 mod mss;
@@ -470,164 +472,6 @@ pub(super) fn lookup_forwarding_resolution_inner_ecmp(
             )
         }
     }
-}
-
-/// #4392: a reject-reply sink for the flow-backed PBR drop path. Present on the
-/// flow-backed session-miss path, which carries a full L4 header and can
-/// synthesize a TCP RST / ICMP-unreachable exactly like a non-PBR `then reject`.
-/// `None` on the flowless path (a non-first fragment / L3-only packet has no L4
-/// header to reflect), where a PBR `reject`/`discard` degrades to a silent drop
-/// — identical to the flowless non-PBR input-filter deny.
-pub(super) struct PbrRejectSink<'a> {
-    pub(super) tx_pipeline: &'a mut crate::afxdp::worker::WorkerTxPipeline,
-    pub(super) ingress_ifindex: i32,
-    pub(super) counters: &'a mut crate::afxdp::BatchCounters,
-}
-
-/// #4392: the route-lookup decision returned by `ingress_route_table_override`.
-///
-/// A PBR term `from { ... } then { routing-instance X; reject | discard; }`
-/// carries BOTH a routing-instance override AND a drop action. Before this fix
-/// the override was applied unconditionally and the packet was FORWARDED into
-/// VRF X — a VRF leak plus a false audit (the filter log recorded a deny while
-/// the data plane forwarded). The drop action now gates the override.
-pub(super) enum RouteOverride {
-    /// No interface input filter affects route lookup here, or no PBR
-    /// routing-instance term matched. Use the default route table.
-    None,
-    /// A PBR routing-instance term matched with a non-drop (accept) action.
-    /// Steer the route lookup to this override table (`<ri>.inet[6].0`) and
-    /// forward — normal policy-based routing, unchanged.
-    Table(String),
-    /// A PBR routing-instance term matched with a `reject`/`discard` action.
-    /// The caller MUST DROP: do NOT apply the override, do NOT route-lookup or
-    /// forward. Any reject reply (TCP RST / ICMP unreachable) has already been
-    /// synthesized inside `ingress_route_table_override` when a `PbrRejectSink`
-    /// was supplied and the action is `reject`; `discard`, and the flowless
-    /// (sink-less) path, drop silently.
-    Drop,
-}
-
-pub(super) fn ingress_route_table_override(
-    forwarding: &ForwardingState,
-    frame: &[u8],
-    meta: UserspaceDpMeta,
-    flow: &SessionFlow,
-    ingress_zone_override: Option<u16>,
-    event_stream: Option<&crate::event_stream::EventStreamWorkerHandle>,
-    now_ns: u64,
-    reject_sink: Option<PbrRejectSink<'_>>,
-) -> RouteOverride {
-    let ingress_ifindex = resolve_ingress_logical_ifindex(
-        forwarding,
-        meta.ingress_ifindex as i32,
-        meta.ingress_vlan_id,
-    )
-    .unwrap_or(meta.ingress_ifindex as i32);
-    let is_v6 = matches!(flow.dst_ip, IpAddr::V6(_));
-    if !crate::filter::interface_filter_affects_route_lookup(
-        &forwarding.filter_state,
-        ingress_ifindex,
-        is_v6,
-    ) {
-        return RouteOverride::None;
-    }
-    // #2362: PBR terms may carry per-packet L4 match conditions (tcp-flags /
-    // is-fragment / icmp-type / icmp-code); build the extra inputs so a
-    // `from { tcp-flags ...; } then routing-instance ...` term matches exactly
-    // the authored packets.
-    let extra = crate::afxdp::frame::term_match_extra_from_frame(frame, meta);
-    let routing_result =
-        match crate::filter::evaluate_interface_filter_routing_instance_event_counted(
-            &forwarding.filter_state,
-            ingress_ifindex,
-            is_v6,
-            flow.src_ip,
-            flow.dst_ip,
-            meta.protocol,
-            flow.forward_key.src_port,
-            flow.forward_key.dst_port,
-            meta.dscp,
-            extra,
-            meta.pkt_len as u64,
-        ) {
-            Some(result) => result,
-            None => return RouteOverride::None,
-        };
-    // #4392: a matched PBR routing-instance term may ALSO carry a drop action
-    // (`then { routing-instance X; reject | discard; }`). Such a term is a DENY,
-    // NOT a forward: the routing-instance override must NOT be applied. On the
-    // flow-backed session-miss path a `PbrRejectSink` is supplied, so a `reject`
-    // synthesizes the TCP RST / ICMP-unreachable reply here — byte-identical to
-    // a non-PBR `then reject` — and its ACTUAL outcome is threaded into the
-    // filter log (#3615) below. A `discard`, and the flowless (sink-less) path,
-    // drop silently.
-    let is_drop = matches!(
-        routing_result.action,
-        crate::filter::FilterAction::Reject | crate::filter::FilterAction::Discard
-    );
-    let reject_reply_enqueued = match (routing_result.action, reject_sink) {
-        (crate::filter::FilterAction::Reject, Some(sink)) => {
-            crate::afxdp::poll_descriptor::reject_reply::enqueue_filter_reject_reply(
-                sink.tx_pipeline,
-                forwarding,
-                sink.ingress_ifindex,
-                frame,
-                meta,
-                flow,
-                sink.counters,
-            )
-        }
-        _ => false,
-    };
-    // #2619: emit the accumulated log_match — it captures fall-through
-    // `then { log; next term; }` terms ahead of the routing-instance term that
-    // the PBR path previously dropped, AND the routing-instance term's own log
-    // (latest matched wins). Its action is already normalized to the verdict the
-    // packet receives (#2616). Falls back to nothing when no matched term logged.
-    if let Some(log_match) = routing_result.log_match {
-        let ingress_zone_id = ingress_zone_override
-            .filter(|id| forwarding.zone_id_to_name.contains_key(id))
-            .or_else(|| forwarding.ifindex_to_zone_id.get(&ingress_ifindex).copied())
-            .or_else(|| {
-                forwarding
-                    .ifindex_to_zone_id
-                    .get(&(meta.ingress_ifindex as i32))
-                    .copied()
-            })
-            .unwrap_or(0);
-        emit_filter_log_event(
-            event_stream,
-            flow,
-            meta,
-            ingress_zone_id,
-            0,
-            log_match.filter_id,
-            log_match.term_id,
-            log_match.action,
-            FilterLogSource::Pbr,
-            // #2520: AppID via the hot-path app_catalog.lookup.
-            resolve_flow_app_id(&forwarding.app_catalog, flow),
-            // #3615/#4392: report the TRUTHFUL reject outcome. A forward
-            // (non-drop) PBR term never rejects (false). A `then reject` on the
-            // flow-backed session-miss path synthesizes an RST/ICMP reply above
-            // and logs REJECT; a `discard`, or the flowless (sink-less) path,
-            // logs the truthful DENY (silent drop).
-            reject_reply_enqueued,
-            now_ns,
-        );
-    }
-    if is_drop {
-        // #4392: reject/discard PBR term — the caller must drop; do NOT apply
-        // the routing-instance override or route-lookup/forward.
-        return RouteOverride::Drop;
-    }
-    let routing_instance = routing_result.routing_instance;
-    RouteOverride::Table(if is_v6 {
-        format!("{routing_instance}.inet6.0")
-    } else {
-        format!("{routing_instance}.inet.0")
-    })
 }
 
 pub(super) fn should_cache_local_delivery_session_on_miss(
